requirements.txt
fastapi>=0.115.0
uvicorn>=0.32.0
httpx>=0.27.0
aiosqlite>=0.20.0
anthropic>=0.40.0
openpyxl>=3.1.0
python-dotenv>=1.0.0
